requirements.txt
yfinance>=0.2.36
pykrx>=1.0.45
python-dotenv>=1.0.0
anthropic>=0.25.0
supabase>=2.4.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
rich>=13.7.0
